// migration-to-ts/src/enum.ts
export enum SourcesTempClass {
  item = '.source__item',
  name = '.source__item-name',
}

export enum NewsTempClass {
  item = '.news__item',
  photo = '.news__meta-photo',
  author = '.news__meta-author',
  date = '.news__meta-date',
  title = '.news__description-title',
  source = '.news__description-source',
  content = '.news__description-content',
  link = '.news__read-more a',
}